Move some gradle logic into buildSrc project
