+++ .gitignore
@@ -15,6 +15,7 @@ project/gen/
 project/.gradle/
 project/build/
 project/app/build/
+project/buildSrc/build
 
 # Local configuration file (sdk path, etc)
 project/local.properties
