+++ project/gradle.properties
@@ -18,8 +18,9 @@ org.gradle.jvmargs=-Xmx4096m -XX:MaxPermSize=512m -XX:+HeapDumpOnOutOfMemoryErro
 org.gradle.parallel=true
 org.gradle.daemon=true
 org.gradle.caching=true
+org.gradle.vfs.watch=true
+org.gradle.configureondemand=true
 
-jacocoVersion = 0.8.7
 android.useAndroidX=true
 android.enableJetifier=true
